+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-06-07
+
+### Fixed
+
+- Fixed an issue causing Item Tracker CSV export to not list all items.
+
 ## 2024-06-06
 
 ### Added